Extract UsageHints to be reused between delegates
- not possible to create composed AccessibilityDelegates in an
isolated fashion (i.e. one that contains multiple others) without
the wrapper one knowing the implementation of the ones it's wrapping
- this is because calling initialize multiple times on the same node
can overwrite information set by the preceding delegate
- it worked for the ActionsAccessiblityDelegate only because the
UsageHintsDelegate was called _after_, and it happened not to
update those actions
- instead, extracted UsageHints and users will have to pass it in
explicitly, apart from UsageHintsAccessibilityDelegate which just
mirrors the API
