+++ core/src/main/java/com/novoda/accessibility/ActionsAccessibilityDelegate.java
@@ -2,7 +2,6 @@ package com.novoda.accessibility;
 
 import android.content.res.Resources;
 import android.os.Bundle;
-import android.support.annotation.StringRes;
 import android.support.v4.view.AccessibilityDelegateCompat;
 import android.support.v4.view.accessibility.AccessibilityNodeInfoCompat;
 import android.view.View;
@@ -11,44 +10,16 @@ public class ActionsAccessibilityDelegate extends AccessibilityDelegateCompat {
 
     private final Resources resources;
     private final Actions actions;
-    private final UsageHintsAccessibilityDelegate usageHintsAccessibilityDelegate;
+    private final UsageHints usageHints;
 
     public ActionsAccessibilityDelegate(Resources resources, Actions actions) {
-        this(resources, actions, new UsageHintsAccessibilityDelegate(resources));
+        this(resources, actions, new UsageHints(resources));
     }
 
-    public ActionsAccessibilityDelegate(Resources resources, Actions actions, UsageHintsAccessibilityDelegate usageHintsAccessibilityDelegate) {
+    public ActionsAccessibilityDelegate(Resources resources, Actions actions, UsageHints usageHints) {
         this.resources = resources;
         this.actions = actions;
-        this.usageHintsAccessibilityDelegate = usageHintsAccessibilityDelegate;
-    }
-
-    /**
-     * Label describing the action that will be performed on click
-     */
-    public void setClickLabel(@StringRes int clickLabel) {
-        usageHintsAccessibilityDelegate.setClickLabel(clickLabel);
-    }
-
-    /**
-     * Label describing the action that will be performed on click
-     */
-    public void setClickLabel(CharSequence clickLabel) {
-        usageHintsAccessibilityDelegate.setClickLabel(clickLabel);
-    }
-
-    /**
-     * Label describing the action that will be performed on long click
-     */
-    public void setLongClickLabel(@StringRes int longClickLabel) {
-        usageHintsAccessibilityDelegate.setLongClickLabel(longClickLabel);
-    }
-
-    /**
-     * Label describing the action that will be performed on long click
-     */
-    public void setLongClickLabel(CharSequence longClickLabel) {
-        usageHintsAccessibilityDelegate.setLongClickLabel(longClickLabel);
+        this.usageHints = usageHints;
     }
 
     @Override
@@ -58,8 +29,7 @@ public class ActionsAccessibilityDelegate extends AccessibilityDelegateCompat {
             String label = resources.getString(action.getLabel());
             info.addAction(new AccessibilityNodeInfoCompat.AccessibilityActionCompat(action.getId(), label));
         }
-
-        usageHintsAccessibilityDelegate.onInitializeAccessibilityNodeInfo(host, info);
+        usageHints.addClickEventUsageHints(host, info);
     }
 
     @Override
