+++ demo/src/main/java/com/novoda/accessibility/demo/custom_actions/TweetView.java
@@ -1,6 +1,7 @@
 package com.novoda.accessibility.demo.custom_actions;
 
 import android.content.Context;
+import android.support.v4.view.AccessibilityDelegateCompat;
 import android.support.v4.view.ViewCompat;
 import android.support.v7.app.AlertDialog;
 import android.util.AttributeSet;
@@ -9,11 +10,7 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.TextView;
-
-import com.novoda.accessibility.AccessibilityServices;
-import com.novoda.accessibility.ActionsMenuAccessibilityDelegate;
-import com.novoda.accessibility.ActionsMenuAlertDialog;
-import com.novoda.accessibility.ActionsMenuInflater;
+import com.novoda.accessibility.*;
 import com.novoda.accessibility.demo.R;
 
 public class TweetView extends LinearLayout {
@@ -45,10 +42,13 @@ public class TweetView extends LinearLayout {
     public void display(String tweet, Listener listener) {
         MenuItem.OnMenuItemClickListener menuItemClickListener = createMenuItemClickListener(tweet, listener);
         Menu menu = actionsMenuInflater.inflate(R.menu.tweet_actions, menuItemClickListener);
-        ActionsMenuAccessibilityDelegate delegate = new ActionsMenuAccessibilityDelegate(menu, menuItemClickListener);
-//      TODO: add composable delegate - delegate.setClickLabel(R.string.tweet_actions_usage_hint);
+        UsageHints usageHints = new UsageHints(getResources());
+        usageHints.setClickLabel(R.string.tweet_actions_usage_hint);
+
+        AccessibilityDelegateCompat delegate = new ActionsMenuAccessibilityDelegate(menu, menuItemClickListener, usageHints);
         ViewCompat.setAccessibilityDelegate(this, delegate);
 
+        setContentDescription(tweet);
         tweetTextView.setText(tweet);
 
         if (services.isSpokenFeedbackEnabled()) {
